requirements.txt
pandas
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
chromadb
google-generativeai
python-dotenv
